Created questions-we-can-ask.md
I created this markdown file to 1. Motivate and remind myself of why I personally want to do this scraping activity, 2. Convey some of the questions that were racing through my mind when discussing the scrapping project and data we wanted to collect and 3. Have a list where we can share questions we want to ask with current or future collaborative scrapping activities.